added module structure to js
