+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "mesto",
+  "homepage": "https://dontstain.github.io/mesto/",
   "version": "1.0.0",
   "description": "Interactive board for cards uploaded by different users. Using object-oriented programming and server requests.\r Using BEM file structure",
   "main": "index.js",
@@ -17,7 +18,6 @@
   "bugs": {
     "url": "https://github.com/dontstain/mesto/issues"
   },
-  "homepage": "https://github.com/dontstain/mesto#readme",
   "devDependencies": {
     "@babel/cli": "^7.7.0",
     "@babel/core": "^7.7.2",
