+++ src/index.html
@@ -6,7 +6,6 @@
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Место</title>
   <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.main.css %>">
-  <script src="./validation.js" defer></script>
   <script src="<%= htmlWebpackPlugin.files.chunks.main.entry %>" defer></script>
 </head>
 <body>
