+++ src/index.js
@@ -1,0 +1,2 @@
+import './pages/style.css';
+import './scripts/script.js';
