+++ webpack.config.js
@@ -7,7 +7,7 @@ const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const isDev = process.env.NODE_ENV === 'development';
 
 module.exports = {
-  entry: { main: './src/scripts/script.js' },
+  entry: { main: './src/index.js' },
   output: {
       path: path.resolve(__dirname, 'dist'),
       filename: '[name].[chunkhash].js'
@@ -43,6 +43,11 @@ module.exports = {
   ]
   },
   plugins: [
+    new HtmlWebpackPlugin({ 
+      inject: false,
+      template: './src/index.html',
+      filename: 'index.html'
+    }),
     new MiniCssExtractPlugin({
       filename: 'style.[contenthash].css',
     }),
@@ -57,11 +62,6 @@ module.exports = {
     new WebpackMd5Hash(),
     new webpack.DefinePlugin({
       'NODE_ENV': JSON.stringify(process.env.NODE_ENV)
-    }),
-    new HtmlWebpackPlugin({ 
-      inject: false,
-      template: './src/index.html',
-      filename: 'index.html'
     })
   ]
 }
